Add SocketIO functionality + more frontend changes
Implements SocketIO functionality (i.e. can handle live feeds and update
the rendering). Also added functionality to deal with multiple subway cars
and the necessary data, as well as variable framerates. Currently does not
perform FPS throttling -- not entirely sure if this is a necessary thing
to consider for our case.