JavaScript style + closing files in Flask
Modified the client code so that the JavaScript conforms to ES6 syntax (more specifically it conforms mainly to AirBnb's JavaScript style). Also added a slight modification for the animation (to avoid possibility of float error accumulation).

Also added context mangers for opening JSON files in the Flask app so that they can be safely closed (was originally hastily written), and encapsulated coordinates into a namedtuple.