Updated README and moved Mapbox GL JS key
Public access token for Mapbox GL JS is now moved to API_KEYS.py (changed from API_KEY.py to reflect this), so that it is not hard-coded in the client code. Key is now passed through Flask when users visit the website and embedded as a global variable.

README also updated to reflect this, plus some formatting fixes + typo fixes.