Refactored JSON script, added graph + cmd line
Did some refactoring of the JSON script so that it can handle command line arguments so that you can choose for it to skip writing certain JSON files.

Instead of a shape_indices.json, the script now writes graph.json. This in particular is a graph structure for the stops in the subway system, which represents stops as vertices and edges represent the possible paths between any two adjacent stops on a particular trip. Thus, they do not necessarily need to be physically adjacent stops -- for example, an express trip may skip intermediate stops, and so those are also accounted for.

The edges represent the sequence of points that represent the path between stops in shapes.txt (represented by the shape ID and the start/end indices). The representation is simply that of the first shape found matching both stops, so the direction may not be correct -- in other words, the first segment seen when constructing that edge may have been southwards in orientation, but the shape found may be northwards in orientation. This is handled appropriately in app.py (see below).

The vertices are mapped by stop IDs and store a dictionary of its adjacent vertices -- the values are +/-1, which represent whether or not the edge between the vertex and its adjacent vertex is the same direction or opposite of the edge orientation (i.e. whether or not it is of the same direction of the first segment seen forming the same edge), since there may be different trips in opposite directions but using the same stops. Thus the sign represents the relative orientation of this particular out-edge and the original directed edge -- this is done to save on storage, since once we have the original out-edge of that particular segment of points, there's no need to store the actual points; we only need to keep track of the relative orientation wrt the first edge. Really, it works if we store relative orientation wrt any particular fixed edge, but the first one seen for that segment is usually the most convenient.

To keep track of both the relative orientation of an out-edge to the first out-edge seen for its segment of points, and also of the orientation of the first out-edge to the actual shape that it's mapped to (as they may also be in opposite orientation), app.py, when implementing get_path, effectively "composes" the two orientations. It stores both orientations as +/-1, which represent 1. relative orientation to original out-edge and 2. orientation of original out-edge to shape. In essence, the signs represent the "flips" or "reversals" required of the actual shape's points to get the actual desired path. Thus, by looking at the sign of the composition, it ensures we only have to perform at most 1 flip.

The reason for creating graph.json is because ideally we would be able to simply get the corresponding shape ID from the trip ID, as the trip path ID is often at the end of the trip ID, and it tends to correspond with the shape ID. But there may be cases in the live feed where it does not correspond with static data (i.e. the live trip path does not correspond with an existing shape ID), a risk mentioned in the MTA's live feed documentation, since there may be maintenance changes or other situations. Thus in these cases there is no way to properly render the subway car if we opted to go for a direct approach and attempt to get the shape ID from the trip ID.

On the other hand, it should be a pretty safe assumption that there is a unique path between two adjacent stops, even for express trips that skip intermediate stops. Even if there is a duplicate path, it shouldn't be an issue, unless one alternative path between two adjacent paths was significantly longer or out of the way -- neither of these cases seem to be the case for the NYC subway system, and the second case doesn't seem particularly likely in general (though I could be wrong). Thus, creating a graph between stops and retrieving the edges seems like a more robust option than relying on knowing the corresponding shape ID for a particular trip, and simply assuming the segment of points between two stops is a unique one.

However, the NYC MTA live feed documentation also mentions that even if there is a partial match for the trip path ID with a shape ID/path in the static data, then it is a real scheduled trip, but the documentation doesn't specify how a partial match determines the exact shape/path in the static data. For example, how should some modified trip like N..N1AR match either N..N16R or N..N19R, which are both static trips? It partially matches both with a difference of 1 character, but no method of tiebreaker specified. Thus, until there's more information on that, the graph option might be safer for now, until there's a flaw pointed out in the assumption of unique/effectively unique paths between adjacent stops on trips.

Will add documentation for code + explanation of some edge cases in script in next commit.