Ability to retrieve previous stop + refactoring
Added PrevStops class; this allows for the app to retrieve the most likely previous stop from a given trip that is in motion (i.e. has a VehiclePosition FeedEntity). Explanation on how this is done is explained in docstrings.

Pickling is now introduced, so that these objects are now cached upon running the parsing methods and can be loaded and used for future runs. static_to_json is now renamed to just static in order to reflect the greater variety of functions that it performs, since it effectively parses through all of the static information to retrieve useful information.

Also did some refactoring; moved some code into classes to abstract away certain operations.

Enhanced the actual parsing so that the parsing can be done without running it from the command line, by simply passing in arguments.

Also added some more documentation + formatting + bug fixes.