Added fix to account for reroutes/service changes
In the case of a service change or reroute, the published route in the live feed may differ from the actual route of stops that it's running. Thus, in order to get the previous stop, some logic was added to account for this. It will first attempt to search within its route group in case something happens like the 3 being routed to the 1 (which tends to be the most likely scenario).

However, there are cases like the 2 running stops for the 5 or vice versa (i.e. outside of a route group), since if no match is found in a route group, get_prev_stops now searches within all routes as a last ditch effort.